third_party/blink/web_tests/platform/generic/fast/forms/select/option-value-and-label-changed-by-js-expected.txt

Test for .value and .label of OPTION element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS o1.value is "newValue"
PASS o1.label is "newLabel"
PASS o2.value is "newValue"
PASS o2.label is "newLabel"
PASS o3.value is "newValue"
PASS o3.label is "newLabel"
PASS o4.value is "newValue"
PASS o4.label is "newLabel"
PASS successfullyParsed is true

TEST COMPLETE

text
text
text
text
